Stock & Crypto API
We have added live data for both Crypto and Stock. We used se Yahoo Finance for stocks (via yfinance library).
Use CoinGecko for crypto (fully free, no API key required). We also included the dictionary for user input so it has to be a real stock or crypto and it'll check before accepting input.